requirements.txt
numpy==2.0.2
pandas==2.3.0
matplotlib==3.9.4
plotly==6.1.2
graphviz==0.21
hdbscan==0.8.39
pytorch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
torchviz==0.0.3
scikit-learn==1.6.1
scipy==1.13.1